refact json parser tests
